# Added an alternative way to generate first-order multiplets with better scaling