@sametz 2024-01-10
Two things:

1. The next version release will be Python 3.8+. That means `math.comb` from the standard library could be used instead of scipy. I'd like to minimize reliance on external libraries where convenient. Could you try implementing your solution with `math.comb`? See [this link](https://stackoverflow.com/questions/26560726/python-binomial-coefficient) on Stack Overflow.

2. How many splittings need there be before your code is faster than the current code? Is there a real-world example that can be used as a benchmark? For hydrocarbons, anything over 8 nuclei would be unusual, but my understanding is that for other spin-half nuclei such as phosphorous more complexity can be seen.